Create 3. Sentiment Analysis.py
This code conduct sentiment analyses on a sentence level, extract the sentences with the highest scores of each attitude and plots the results. 
It uses the BERT Tone model to classify polarity and tone with the respective classes negative, neutral, positive and subjective, objective. The BERT emotion is used to classify Expectation, No Emotion, Contempt, Trust, Fear, Anger, Surprise, Joy, Sadness.